Switch to a pure git managed dotfiles project
Drops the reliance on symlinks and a script to synchronize src to $HOME.
This also means that we don't have to worry about intermediary
directories, stale symlinks, or the synch script itself.

It works with a bare git repository and an aliased git that uses the
bare repo as the git directory and $HOME as the working tree. Details in
the README.

It requires an initial bootstrap checkout, and can run into conflicts
with pre-existing dotfiles.

https://developer.atlassian.com/blog/2016/02/best-way-to-store-dotfiles-git-bare-repo/