(maint) Remove -c pane path from split-window bindings
This was there to ensure that a split-window command maintained the cwd
of the window it was split from. But this wasn't working in centos. The
functionality is built into recent version of tmux (1.7?) and the
workaround no longer seems to be needed for Ubuntu either. Removing the
-c arg lets split work as expected for carrying along cwd on both Ubuntu
and Centos.