Revert "(maint) Remove -c pane path from split-window bindings"
This reverts commit 878a400fc8151e29481ed6755e95cb37e60a8aba.

I was wrong about -c "#{pane_current_path}". That is the correct setting
to get split panes to keep the directory of the previously focused pane,
but the el7 tmux is ancient (1.8) and doesn't support it correctly.

For those versions, not having -c "#{pane_current_path}" allows split to
work they I expect.